# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactPreview": {
        "enum": [
          "auto",
          "text",
          "json",
          "markdown",
          "image",
          "html",
          "pdf",
          "audio",
          "video",
          "none"
        ],
        "type": "string"
      },
...
      "ConfirmSpec": {
        "properties": {
          "details": {
            "default": null
          },
          "dry_run": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "operation_digest": {
            "default": "",
            "description": "Canonical digest of target + details binding this approval to the\nexact operation content (A06). Approvals never transfer across\ndifferent digests even for the same node and kind.",
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "kind",
          "target",
          "dry_run"
        ],
        "type": "object"
      },
...
      "FormSpec": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/InputField"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "title_i18n": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "id",
          "title",
          "fields"
        ],
        "type": "object"
      },
...
      "InputField": {
        "additionalProperties": false,
        "properties": {
          "default": {
            "default": null
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "description_i18n": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "item_type": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "label_i18n": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "min_items": {
            "default": null,
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "option_labels_i18n": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "object"
          },
          "options": {
            "default": [],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "pattern": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "placeholder": {
            "type": [
              "string",
              "null"
            ]
          },
          "placeholder_i18n": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "required": {
            "default": false,
            "type": "boolean"
          },
          "schema": {
            "description": "Optional JSON Schema applied after the canonical field-type checks."
          },
          "type": {
            "type": "string"
          },
          "ui": {
            "additionalProperties": true,
            "description": "Renderer-specific presentation metadata forwarded to clients unchanged.",
            "type": "object"
          }
        },
        "required": [
          "id",
          "type"
        ],
        "type": "object"
      },
...
      "McpServerList": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/McpServerSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "McpServerList",
        "type": "object"
      },
      "McpServerSummary": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "auth": {
            "type": "string"
          },
          "authorized": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "transport": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "transport",
          "auth",
          "authorized"
        ],
        "title": "McpServerSummary",
        "type": "object"
      },
...
      "OutputArtifact": {
        "properties": {
          "bytes": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "description": {
            "default": "",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "mime": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          },
          "preview": {
            "$ref": "#/components/schemas/ArtifactPreview",
            "default": "auto"
          },
          "required": {
            "type": "boolean"
          },
          "sha256": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "sha256",
          "bytes",
          "label",
          "required"
        ],
        "type": "object"
      },
      "OutputManifest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/OutputArtifact"
            },
            "type": "array"
          }
        },
        "required": [
          "artifacts"
        ],
        "title": "OutputManifest",
        "type": "object"
      },
...
      "ProblemDetails": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemFieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "type": "object"
      },
...
      "RunListItem": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "generator_id": {
            "type": "string"
          },
          "run_id": {
            "type": "string"
          },
          "seq": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "started_at": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "state",
          "generator_id",
          "started_at"
        ],
        "title": "RunListItem",
        "type": "object"
      },
      "RunListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "type": "array"
          },
          "next_cursor": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "title": "RunListResponse",
        "type": "object"
      },
      "RunMetrics": {
        "properties": {
          "cost_microusd": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "duration_ms": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "llm_calls": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "regenerate_attempts": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "repair_attempts": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "steps_executed": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "steps_failed": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "steps_skipped": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "steps_succeeded": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "steps_total": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "tokens_cached_input": {
            "default": 0,
            "description": "Input tokens served from provider cache, included in `tokens_input`.",
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "tokens_input": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "tokens_output": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "tokens_total": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "RunSnapshot": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "artifacts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputManifest"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_sha256": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "generator_id": {
            "description": "Owning generator id so clients never parse it from run_id (C03).\nRequired: a snapshot without an owner is corrupt, never something\nto guess from the run id.",
            "type": "string"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunMetrics"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Cost totals: exact terminal metrics when finished, otherwise a live\nsynthesis from the folded budget. Absent only before any activity."
          },
          "parent_run_id": {
            "default": null,
            "description": "Parent run id for runs created by fork.",
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "default": 0,
            "description": "Scheduling priority, higher runs first.",
            "format": "int32",
            "type": "integer"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FormSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "queue_position": {
            "default": null,
            "description": "1-based position among queued runs, present only while `Queued`.",
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "queued_at": {
            "default": null,
            "description": "RFC 3339 timestamp of the last transition to `Queued`, if any.",
            "type": [
              "string",
              "null"
            ]
          },
          "run_id": {
            "type": "string"
          },
          "seq": {
            "default": 0,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "state",
          "generator_id"
        ],
        "title": "RunSnapshot",
        "type": "object"
      },
...
      "RunStatus": {
        "oneOf": [
          {
            "enum": [
              "queued",
              "running",
              "waiting",
              "confirming",
              "succeeded",
              "failed",
              "canceled",
              "interrupted"
            ],
            "type": "string"
          },
          {
            "const": "cancel_requested",
            "description": "A cancel request was accepted into the mailbox but no terminal\noutcome is journaled yet. This is acceptance, not settlement: only\na journaled terminal state may report `Canceled` (A02).",
            "type": "string"
          }
        ]
      },
...
      "StartRun": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "answers": {
            "additionalProperties": true,
            "default": {},
            "description": "Pre-provisioned answers keyed by question id, consumed without interaction.\nSame semantics as eval suite answers.",
            "type": "object"
          },
          "confirmations": {
            "additionalProperties": {
              "type": "boolean"
            },
            "default": {},
            "description": "Pre-provisioned confirmation decisions keyed by confirmation id.",
            "type": "object"
          },
          "generator_id": {
            "type": "string"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object"
          },
          "priority": {
            "default": null,
            "description": "Scheduling priority, higher runs first. Defaults to 0.",
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "generator_id",
          "inputs"
        ],
        "title": "StartRun",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "qcg",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/mcp/servers": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerList"
                }
              }
            },
            "description": "Configured MCP servers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List configured MCP servers and authorization status"
      }
    },
...
    "/api/runs": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RunStatus"
            }
          },
          {
            "in": "query",
            "name": "generator_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Known runs"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List runs"
      },
      "post": {
        "parameters": [
          {
            "description": "Retries with the same key and request body return the original run.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRun"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            },
            "description": "Started run",
            "headers": {
              "Location": {
                "description": "URL of the newly created run.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Payload too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Start a run"
      }
    },
    "/api/runs/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Run deleted"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Delete a terminal run"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Return 304 when this validator matches the current representation.",
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            },
            "description": "Run snapshot",
            "headers": {
              "ETag": {
                "description": "Weak validator for conditional GET requests.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified",
            "headers": {
              "ETag": {
                "description": "Weak validator for conditional GET requests.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Run snapshot"
      }
    },
...
  },
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}